Choose the XSL file based on the report version
Starting with cpptest 9.0, the report file format changed slightly,
fortunately the tool version can be found at the same address xml-wise.

This implementation uses a stream reader to avoid completely reading the
report twice: once for the version, and once for the conversion. This is
convenient since the tool version is in the very first element.

The file `cpptestunit-9.0-to-junit-1.0.xsl` isn't actually part of this
commit.

refs JENKINS-18727