Use a Digester instead of StAX to get the toolVer
The choice of StAX was made because the streaming aspect allowed a small
memory footprint. The `toolVer` attribute is located in the very first
XML tag, so reading the whole report was not needed. The issue is that
it failed on a slave. It now uses a `Digester` from commons-digester.

The version extraction has been moved to a new VersionParser, also only
visible within the same package. Moving to a dedicated class made it
easier to test this very part.