Refactoring of the parsing code
The main idea is to separate the parsing "plumbing" (commons-digester)
from the data model. Instead of an anemic object model, the logic has
been moved to the mapped classes.

A `FileAnnotationBuilder` interface with package visibility was added to
create a bridge between the XML parsing and the data model. Overall, it
means less lines of code, and less complexity in various places.

Unused fields were also removed from the object model. They can be added
back when needed, but right now, YAGNI...